+++ crates/bcr-ebill-api/src/service/company_service.rs
@@ -145,18 +145,15 @@ impl CompanyService {
         public_key: &str,
     ) -> Result<Option<File>> {
         if let Some(upload_id) = upload_id {
-            let files = self
+            let (file_name, file_bytes) = &self
                 .file_upload_store
-                .read_temp_upload_files(upload_id)
+                .read_temp_upload_file(upload_id)
                 .await
                 .map_err(|_| crate::service::Error::NoFileForFileUploadId)?;
-            if !files.is_empty() {
-                let (file_name, file_bytes) = &files[0];
-                let file = self
-                    .encrypt_and_save_uploaded_file(file_name, file_bytes, id, public_key)
-                    .await?;
-                return Ok(Some(file));
-            }
+            let file = self
+                .encrypt_and_save_uploaded_file(file_name, file_bytes, id, public_key)
+                .await?;
+            return Ok(Some(file));
         }
         Ok(None)
     }
@@ -907,13 +904,8 @@ pub mod tests {
             })
         });
         file_upload_store
-            .expect_read_temp_upload_files()
-            .returning(|_| {
-                Ok(vec![(
-                    "some_file".to_string(),
-                    "hello_world".as_bytes().to_vec(),
-                )])
-            });
+            .expect_read_temp_upload_file()
+            .returning(|_| Ok(("some_file".to_string(), "hello_world".as_bytes().to_vec())));
         file_upload_store
             .expect_remove_temp_upload_folder()
             .returning(|_| Ok(()));
@@ -1062,13 +1054,8 @@ pub mod tests {
             })
         });
         file_upload_store
-            .expect_read_temp_upload_files()
-            .returning(|_| {
-                Ok(vec![(
-                    "some_file".to_string(),
-                    "hello_world".as_bytes().to_vec(),
-                )])
-            });
+            .expect_read_temp_upload_file()
+            .returning(|_| Ok(("some_file".to_string(), "hello_world".as_bytes().to_vec())));
         file_upload_store
             .expect_remove_temp_upload_folder()
             .returning(|_| Ok(()));