+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # 0.3.1
 
+* Added `log_level` to Config, which defaults to `info`
+* Changed the API for uploading files to bill to use `file` instead of `files`.
+So files can only be uploaded individually, but for `issue()`, `file_upload_ids`
+can be passed - a list of file upload ids, to upload multiple files for one bill.
 * Restructured `BitcreditBillWeb` to a more structured approach, separating `status`,
 `data` and `participants` and adding the concept of `current_waiting_state`, to
 have all data available, if the bill is in a waiting state.